Added workaround to fix a bug where the programm wasn't running when invoked multiple times.
The current "solution" is pretty memory hungry, as it returns the entire data from the detectors and for the projected templates.
Only later the SNR gets calculated.
Before the SNR was calculated directly when creating the templates. Hence the memory use was minimal. This however caused
problems when invoking the function more than once. It just got stuck at trying to calculate the matched_filter function.
Having the programm run in serial and it working made clear, that the problem lies within the matched_filter function.
Until a solution to this problem is found the programm is mostly unusable, as 1000 templates already fill up 20GB of memory.