Made verify_templates.py more useful.
This file now provides functionality, which is not yet implemented in the main program. It will be used to check wether or not the template-file that is being loaded is conformal
with the requirements set in the options. It also provides a boolean function now which returns True if the calculated SNR roughly estimates the label SNR. This one however does
not yet check if the mean value of the difference is close to 0.