Worked out the bug present in previous parts and implemented the use of .ini files.
Nothing in the source code had to be changed in order to fix the bug where the program would just stop at some point. We just needed to delete the template files previously created
and run the command OMP_NUM_THREADS=1 python <file-name>.
The .ini file is now used to set the default values that have to be filtered out during this stage.