Added support for .ini files as an option argument.
The function checks wether or not an option 'ini_file' was given and if it is a valid file when given an argument.
When the file actually exists, it loads all the parameters specified in the .ini file and uses them as input-options.
Each of these options is overwritten in case the function is called with the same argument that has a different value.