Indroduced two new options.
For run_net there is now an option to use a custom compilation function for the network. This has to be specified in the file were the
network is specified.

For generating templates there now is an option to set a default SNR when there are is no GW in the data.