Fixed bug and implemented new functions.
The bug that was fixed was an issue coming from certain characters
that TeX cannot display so easily. These were changed to be
displayable.
The new functions are the old functions but without the
requirement of providing a data-object, as that information should
now always be stored in the prediction file. Also the need for the
false alarm rate path in the sensitivity plot was dropped in these
new functions.
evaluate_nets.evaluate_training might at some point switch to
using these new functions.