Changed training function.
The function now names the savefiles as the following evaluating
functions expect the files to be named. It also now correctly
stores the loss in the results file and stores another file,
that holds not only the loss values but also the named keys to
them, in order to allow accurate reconstruction.